fix #6 - fix broken test, mark dist DEPRECATED
Swagger2 has been DEPRECATED so this distribution follows, the fix
here is to skip the test as i can't see a way to fix it without
horrible hacks or changing the expected API Blueprint output - the
reason the test fails is the validator now resolved references, so
there is no way for us to know what the resolved reference is (in
terms of the schema name, Coupon in this case) so we can't make the
output the same. We could just print out the object structure but
we lose the referential aspect that API Blueprint is supposed to
take. Maybe this doesn't matter so much but i have little interest
in fixing this.